less parrallel threads for tnx0.25v4; store annual ts files

+++ packages/MICOM_DIAG/code/compute_ann_time_series.sh
@@ -42,9 +42,15 @@ var_list=`cat $WKDIR/attributes/vars_ts_ann_${casename}_${filetype}`
 first_yr_prnt=`printf "%04d" ${first_yr}`
 last_yr_prnt=`printf "%04d" ${last_yr}`
 ann_ts_file=${casename}_ANN_${first_yr_prnt}-${last_yr_prnt}_ts_${filetype}.nc
+ann_ts_var_list="mmflxd voltr temp saln templvl salnlvl"
 
 # Calculate number of chunks and the residual
-nproc=10
+if [ `cat $WKDIR/attributes/grid_${casename}` == "tnx0.25v4" ]
+then
+    nproc=5
+else
+    nproc=10
+fi
 let "nyrs = $last_yr - $first_yr + 1"
 let "nchunks = $nyrs / $nproc"
 let "residual = $nyrs % $nproc"
@@ -90,14 +96,30 @@ do
             let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
             yr_prnt=`printf "%04d" ${YR}`
             filename=${casename}.micom.hy.${yr_prnt}.nc
-            eval $NCKS -O -v $var_list --no_tmp_fl $pathdat/$filename $WKDIR/${casename}_ANN_${yr_prnt}.nc &
-            pid+=($!)
+            fflag=1     #check if all required annual ts files exist
+            ann_ts_var_list="mmflxd voltr temp saln templvl salnlvl"
+            for var in `echo $var_list | sed 's/,/ /g'` ; do
+                tsfile=${var}_${casename}_ANN_${filetype}_${yr_prnt}.nc
+                if [ "${ann_ts_var_list/$var}" != "${ann_ts_var_list}" ] && [ ! -f  $tsdir/ann_ts/${tsfile} ]; then
+                    echo $tsdir/ann_ts/${tsfile}
+                    fflag=0
+                    break
+                fi
+            done
+            if [ $fflag  = 0 ]; then
+                eval $NCKS -O -v $var_list --no_tmp_fl $pathdat/$filename $WKDIR/${casename}_ANN_${yr_prnt}.nc &
+                pid+=($!)
+            else
+                echo Skip computing year ${yr_prnt}, time-series variables already exist.
+            fi
             let iproc++
         done
-        for ((m=0;m<=$nyrsm;m++))
+        for ((m=0;m<${#pid[*]};m++))
         do
             wait ${pid[$m]}
             if [ $? -ne 0 ]; then
+                let "YR = ($ichunk - 1) * $nproc + $m + $first_yr"
+                yr_prnt=`printf "%04d" ${YR}`
                 echo "ERROR in extracting variables from annual history file: $NCKS -O -v $var_list --no_tmp_fl $pathdat/$filename $WKDIR/${casename}_ANN_${yr_prnt}.nc"
                 echo "*** EXITING THE SCRIPT ***"
                 exit 1
@@ -119,14 +141,29 @@ do
                 filename=${casename}.micom.hm.${yr_prnt}-${mon}.nc
                 filenames+=($filename)
             done
-            eval $NCRA -O --no_tmp_fl --hdr_pad=10000 -w 31,28,31,30,31,30,31,31,30,31,30,31 -v $var_list -p $pathdat ${filenames[*]} $WKDIR/${casename}_ANN_${yr_prnt}.nc &
-            pid+=($!)
+            fflag=1     #check if all required annual ts files exist
+            for var in `echo $var_list | sed 's/,/ /g'` ; do
+                tsfile=${var}_${casename}_ANN_${filetype}_${yr_prnt}.nc
+                if [ "${ann_ts_var_list/$var}" != "${ann_ts_var_list}" ] && [ ! -f $tsdir/ann_ts/${tsfile} ]; then
+                    echo $tsdir/ann_ts/${tsfile}
+                    fflag=0
+                    break
+                fi
+            done
+            if [ $fflag = 0 ]; then
+                eval $NCRA -O --no_tmp_fl --hdr_pad=10000 -w 31,28,31,30,31,30,31,31,30,31,30,31 -v $var_list -p $pathdat ${filenames[*]} $WKDIR/${casename}_ANN_${yr_prnt}.nc &
+                pid+=($!)
+            else
+                echo Skip computing year ${yr_prnt}, time-series variables already exist.
+            fi
             let iproc++
         done
-        for ((m=0;m<=$nyrsm;m++))
+        for ((m=0;m<${#pid[*]};m++))
         do
             wait ${pid[$m]}
             if [ $? -ne 0 ]; then
+                let "YR = ($ichunk - 1) * $nproc + $m + $first_yr"
+                yr_prnt=`printf "%04d" ${YR}`
                 echo "ERROR in computing annual means from monthly history files: $NCRA -O --no_tmp_fl --hdr_pad=10000 -w 31,28,31,30,31,30,31,31,30,31,30,31 -v $var_list -p $pathdat $filenames $WKDIR/${casename}_ANN_${yr_prnt}.nc"
                 echo "*** EXITING THE SCRIPT ***"
                 exit 1
@@ -142,12 +179,14 @@ do
         let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
         yr_prnt=`printf "%04d" ${YR}`
         filename=${casename}_ANN_${yr_prnt}.nc
-        $NCKS --quiet -d depth,0 -d x,0 -d y,0 -v parea $WKDIR/$filename >/dev/null 2>&1
-        if [ $? -ne 0 ]; then
-            $NCKS --quiet -A -v parea -o $WKDIR/$filename $grid_file
-            $NCAP2 -O -s 'dmass=dp*parea' $WKDIR/$filename  -o dmass.nc >/dev/null 2>&1
-            $NCKS --quiet -A -v dmass -o $WKDIR/$filename dmass.nc >/dev/null 2>&1
-            rm -f dmass.nc
+        if [ -f $WKDIR/$filename ]; then
+            $NCKS --quiet -d depth,0 -d x,0 -d y,0 -v parea $WKDIR/$filename >/dev/null 2>&1
+            if [ $? -ne 0 ]; then
+                $NCKS --quiet -A -v parea -o $WKDIR/$filename $grid_file
+                $NCAP2 -O -s 'dmass=dp*parea' $WKDIR/$filename  -o dmass.nc >/dev/null 2>&1
+                $NCKS --quiet -h -A -v dmass -o $WKDIR/$filename dmass.nc >/dev/null 2>&1
+                rm -f dmass.nc
+            fi
         fi
         let iproc++
     done
@@ -164,15 +203,19 @@ do
                 let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
                 yr_prnt=`printf "%04d" ${YR}`
                 infile=${casename}_ANN_${yr_prnt}.nc
-                outfile=${var}_${casename}_ANN_${yr_prnt}.nc
-                eval $NCWA --no_tmp_fl -O -v $var -w dmass -a sigma,y,x $WKDIR/$infile $WKDIR/$outfile &
-                pid+=($!)
+                outfile=${var}_${casename}_ANN_${filetype}_${yr_prnt}.nc
+                if [ ! -f $tsdir/ann_ts/$outfile ]; then
+                    eval $NCWA --no_tmp_fl -O -v $var -w dmass -a sigma,y,x $WKDIR/$infile $WKDIR/$outfile &
+                    pid+=($!)
+                fi
                 let iproc++
             done
-            for ((m=0;m<=$nyrsm;m++))
+            for ((m=0;m<${#pid[*]};m++))
             do
                 wait ${pid[$m]}
                 if [ $? -ne 0 ]; then
+                    let "YR = ($ichunk - 1) * $nproc + $m + $first_yr"
+                    yr_prnt=`printf "%04d" ${YR}`
                     echo "ERROR in calculating mass weighted global average: $NCWA --no_tmp_fl -O -v $var -w dmass -a sigma,y,x $WKDIR/$infile $WKDIR/$outfile"
                     echo "*** EXITING THE SCRIPT ***"
                     exit 1
@@ -190,15 +233,19 @@ do
                 let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
                 yr_prnt=`printf "%04d" ${YR}`
                 infile=${casename}_ANN_${yr_prnt}.nc
-                outfile=${var}_${casename}_ANN_${yr_prnt}.nc
-                eval $NCWA --no_tmp_fl -O -v $var -w parea -a x,y $WKDIR/$infile $WKDIR/$outfile &
-                pid+=($!)
+                outfile=${var}_${casename}_ANN_${filetype}_${yr_prnt}.nc
+                if [ ! -f $tsdir/ann_ts/$outfile ]; then
+                    eval $NCWA --no_tmp_fl -O -v $var -w parea -a x,y $WKDIR/$infile $WKDIR/$outfile &
+                    pid+=($!)
+                fi
                 let iproc++
             done
-            for ((m=0;m<=$nyrsm;m++))
+            for ((m=0;m<${#pid[*]};m++))
             do
                 wait ${pid[$m]}
                 if [ $? -ne 0 ]; then
+                    let "YR = ($ichunk - 1) * $nproc + $m + $first_yr"
+                    yr_prnt=`printf "%04d" ${YR}`
                     echo "ERROR in calculating area weighted global average: $NCWA --no_tmp_fl -O -v $var -w parea -a x,y $WKDIR/$infile $WKDIR/$outfile"
                     echo "*** EXITING THE SCRIPT ***"
                     exit 1
@@ -215,21 +262,23 @@ do
                 let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
                 yr_prnt=`printf "%04d" ${YR}`
                 infile=${casename}_ANN_${yr_prnt}.nc
-                outfile_tmp=${var}_${casename}_ANN_${yr_prnt}_tmp.nc
-                     outfile=${var}_${casename}_ANN_${yr_prnt}.nc
-                # Max AMOC 20-60N
-                $NCKS -F --no_tmp_fl -O -v $var -d lat,20.0,60.0 -d region,1 $WKDIR/$infile $WKDIR/$outfile_tmp
-                $NCAP2 -O -s 'mmflxd_max=mmflxd.max($lat,$depth)' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp
-                $NCKS --no_tmp_fl -O -v mmflxd_max,region $WKDIR/$outfile_tmp $WKDIR/$outfile
-                # Max AMOC at 26.5N
-                $NCKS -F --no_tmp_fl -O -v $var -d lat,26.5 -d region,1 $WKDIR/$infile $WKDIR/$outfile_tmp
-                $NCAP2 -O -s 'mmflxd265=mmflxd.max($lat,$depth)' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp
-                $NCKS --no_tmp_fl -A -v mmflxd265 -o $WKDIR/$outfile $WKDIR/$outfile_tmp
-                # Max AMOC at 45N
-                $NCKS -F --no_tmp_fl -O -v $var -d lat,45.0 -d region,1 $WKDIR/$infile $WKDIR/$outfile_tmp
-                $NCAP2 -O -s 'mmflxd45=mmflxd.max($lat,$depth)' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp
-                $NCKS --no_tmp_fl -A -v mmflxd45 -o $WKDIR/$outfile $WKDIR/$outfile_tmp
-                rm -f $WKDIR/$outfile_tmp
+                outfile_tmp=${var}_${casename}_ANN_${filetype}_${yr_prnt}_tmp.nc
+                outfile=${var}_${casename}_ANN_${filetype}_${yr_prnt}.nc
+                if [ ! -f $tsdir/ann_ts/$outfile ]; then
+                    # Max AMOC 20-60N
+                    $NCKS -F --no_tmp_fl -O -v $var -d lat,20.0,60.0 -d region,1 $WKDIR/$infile $WKDIR/$outfile_tmp
+                    $NCAP2 -O -s 'mmflxd_max=mmflxd.max($lat,$depth)' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp
+                    $NCKS --no_tmp_fl -O -v mmflxd_max,region $WKDIR/$outfile_tmp $WKDIR/$outfile
+                    # Max AMOC at 26.5N
+                    $NCKS -F --no_tmp_fl -O -v $var -d lat,26.5 -d region,1 $WKDIR/$infile $WKDIR/$outfile_tmp
+                    $NCAP2 -O -s 'mmflxd265=mmflxd.max($lat,$depth)' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp
+                    $NCKS --no_tmp_fl -A -v mmflxd265 -o $WKDIR/$outfile $WKDIR/$outfile_tmp
+                    # Max AMOC at 45N
+                    $NCKS -F --no_tmp_fl -O -v $var -d lat,45.0 -d region,1 $WKDIR/$infile $WKDIR/$outfile_tmp
+                    $NCAP2 -O -s 'mmflxd45=mmflxd.max($lat,$depth)' $WKDIR/$outfile_tmp $WKDIR/$outfile_tmp
+                    $NCKS --no_tmp_fl -A -v mmflxd45 -o $WKDIR/$outfile $WKDIR/$outfile_tmp
+                    rm -f $WKDIR/$outfile_tmp
+                fi
                 let iproc++
             done
         fi
@@ -242,8 +291,10 @@ do
                 let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
                 yr_prnt=`printf "%04d" ${YR}`
                 infile=${casename}_ANN_${yr_prnt}.nc
-                     outfile=${var}_${casename}_ANN_${yr_prnt}.nc
-                $NCKS --no_tmp_fl -O -v voltr,section $WKDIR/$infile $WKDIR/$outfile
+                outfile=${var}_${casename}_ANN_${filetype}_${yr_prnt}.nc
+                if [ ! -f $tsdir/ann_ts/$outfile ]; then
+                    $NCKS --no_tmp_fl -O -v voltr,section $WKDIR/$infile $WKDIR/$outfile
+                fi
                 let iproc++
             done
         fi
@@ -255,29 +306,36 @@ do
         let "YR = ($ichunk - 1) * $nproc + $iproc + $first_yr - 1"
         yr_prnt=`printf "%04d" ${YR}`
         filename=${casename}_ANN_${yr_prnt}.nc
-        rm -f $WKDIR/$filename
+        if [ -f  $WKDIR/$filename ]; then
+            rm -f $WKDIR/$filename
+        fi
         let iproc++
     done
     let ichunk++
 done
 
 # Concancate files
+if [ ! -d $tsdir/ann_ts ]; then
+    mkdir -p $tsdir/ann_ts
+fi
+let "nyrs = $last_yr - $first_yr + 1"
 first_var=1
 for var in `echo $var_list | sed 's/,/ /g'`
 do
-    first_file=${var}_${casename}_ANN_${first_yr_prnt}.nc
-    if [ -f $WKDIR/$first_file ]; then
+    mv $WKDIR/${var}_${casename}_ANN_${filetype}_*.nc $tsdir/ann_ts/ >/dev/null 2>&1
+    first_file=${var}_${casename}_ANN_${filetype}_${first_yr_prnt}.nc
+    if [ -f $tsdir/ann_ts/$first_file ]; then
         echo "Merging all $var time series files..."
-        $NCRCAT --no_tmp_fl -O $WKDIR/${var}_${casename}_ANN_????.nc $WKDIR/${var}_${casename}_ANN_${first_yr_prnt}-${last_yr_prnt}.nc
+        $NCRCAT --no_tmp_fl -O -p $tsdir/ann_ts -n ${nyrs},4,1 ${var}_${casename}_ANN_${filetype}_${first_yr_prnt}.nc -o $WKDIR/${var}_${casename}_ANN_${filetype}_${first_yr_prnt}-${last_yr_prnt}.nc
         if [ $? -eq 0 ]; then
             if [ $first_var -eq 1 ]; then
                 first_var=0
-                  mv $WKDIR/${var}_${casename}_ANN_${first_yr_prnt}-${last_yr_prnt}.nc $tsdir/$ann_ts_file
+                mv $WKDIR/${var}_${casename}_ANN_${filetype}_${first_yr_prnt}-${last_yr_prnt}.nc $tsdir/$ann_ts_file
             else
-                $NCKS -A -o $tsdir/$ann_ts_file $WKDIR/${var}_${casename}_ANN_${first_yr_prnt}-${last_yr_prnt}.nc
+                $NCKS -A -o $tsdir/$ann_ts_file $WKDIR/${var}_${casename}_ANN_${filetype}_${first_yr_prnt}-${last_yr_prnt}.nc
+                rm -f $WKDIR/${var}_${casename}_ANN_${filetype}_${first_yr_prnt}-${last_yr_prnt}.nc
             fi
         fi
-        rm -f $WKDIR/${var}_${casename}_ANN_*.nc
     fi
 done
